Improve CustomServices feature tests
