+++ src/Framework/CustomServiceAwareTrait.php
@@ -9,17 +9,17 @@ use Gacela\Framework\ClassResolver\CustomService\CustomServiceResolver;
 trait CustomServiceAwareTrait
 {
     /** @var array<string,AbstractCustomService> */
-    private array $flexibleServices = [];
+    private array $customServices = [];
 
     public function __call(string $name, array $arguments = []): AbstractCustomService
     {
         $resolvableType = ltrim($name, 'get');
 
-        if (!isset($this->flexibleServices[$resolvableType])) {
-            $this->flexibleServices[$resolvableType] = (new CustomServiceResolver($resolvableType))
+        if (!isset($this->customServices[$resolvableType])) {
+            $this->customServices[$resolvableType] = (new CustomServiceResolver($resolvableType))
                 ->resolve($this);
         }
 
-        return $this->flexibleServices[$resolvableType];
+        return $this->customServices[$resolvableType];
     }
 }
