--- tests/Feature/Framework/CustomServiceOnFactory/CustomModule/Config.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Feature\Framework\CustomServiceOnFactory\CustomModule;
-
-use Gacela\Framework\AbstractConfig;
-
-final class Config extends AbstractConfig
-{
-    /**
-     * @return array<string,int>
-     */
-    public function getAllKeyValues(): array
-    {
-        return [
-            'from-config' => (int)$this->get('from-config'),
-        ];
-    }
-}
