--- tests/Feature/Framework/CustomServiceOnFactory/CustomModule/Facade.php
@@ -1,21 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Feature\Framework\CustomServiceOnFactory\CustomModule;
-
-use Gacela\Framework\AbstractFacade;
-
-/**
- * @method Factory getFactory()
- */
-final class Facade extends AbstractFacade
-{
-    /**
-     * @return array<string,int>
-     */
-    public function findAllKeyValuesUsingRepository(): array
-    {
-        return $this->getFactory()->findAllKeyValuesUsingRepositoryFromFactory();
-    }
-}
