--- tests/Feature/Framework/CustomServiceOnFactory/CustomModule/Infrastructure/Factory.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Feature\Framework\CustomServiceOnFactory\CustomModule\Infrastructure;
-
-use Gacela\Framework\AbstractFactory;
-
-final class Factory extends AbstractFactory
-{
-    /**
-     * @return array<string,int>
-     */
-    public function createDummyArray(): array
-    {
-        return [
-            'from-factory' => 1,
-        ];
-    }
-}
