--- tests/Feature/Framework/CustomServiceOnFactory/config/default.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-return [
-    'from-config' => 1,
-];
