+++ tests/Feature/Framework/CustomServices/CustomModule/Infrastructure/Factory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Feature\Framework\CustomServiceOnFacade\CustomModule\Infrastructure;
+namespace GacelaTest\Feature\Framework\CustomServices\CustomModule\Infrastructure;
 
 use Gacela\Framework\AbstractFactory;
